Don't write cache out again if it didn't change

+++ scraper/scraper/HTMLCache.ts
@@ -3,13 +3,16 @@ import zlib from 'zlib';
 
 export class HTMLCache {
   private data: Map<string, string>;
+  private touched: boolean;
 
   constructor () {
     this.data = new Map();
+    this.touched = false;
   }
 
   set (key: string, value: string): void {
     this.data.set(key, value);
+    this.touched = true;
   }
 
   has (key: string): boolean {
@@ -20,11 +23,13 @@ export class HTMLCache {
     return this.data.get(key);
   }
 
-  async write (destination: string) {
-    const arrayData = Array.from(this.data);
-    arrayData.sort((a, b) => +(a > b) - +(a < b));
-    const json = zlib.brotliCompressSync(Buffer.from(JSON.stringify(arrayData)));
-    writeFileSync(destination, json, 'utf-8');
+  async write (destination: string, force = false) {
+    if (this.touched || force) {
+      const arrayData = Array.from(this.data);
+      arrayData.sort((a, b) => +(a > b) - +(a < b));
+      const json = zlib.brotliCompressSync(Buffer.from(JSON.stringify(arrayData)));
+      writeFileSync(destination, json, 'utf-8');
+    }
   }
 
   async load (source: string) {
